+++ pyne/mesh.py
@@ -1571,6 +1571,200 @@ class StatMesh(Mesh):
 
         return mesh_1
 
+class MeshTally(StatMesh):
+    """This class stores all information from all single mesh tally that
+    exists within some meshtal or state point file. Header information is
+    stored as attributes and the "mesh" attribute is a MOAB mesh with all
+    result and relative error data tagged. This class inherits from StatMesh,
+    exposing all statistical mesh manipulation methods.
+
+    Attributes
+    ----------
+    tally_number : int
+        The tally number.
+        For mesh tally from MCNP, it must end with 4 (e.g. 4, 14, 214).
+        For mesh tally from OpenMC, it could be any int.
+    particle : string
+        Either "neutron" for a neutron mesh tally or "photon" for a photon mesh
+        tally.
+    dose_response : bool
+        True is the tally is modified by a dose response function.
+    x_bounds : list of floats
+        The locations of mesh vertices in the x direction.
+    y_bounds : list of floats
+        The locations of mesh vertices in the y direction.
+    z_bounds : list of floats
+        The locations of mesh vertices in the z direction.
+    e_bounds : list of floats
+        The minimum and maximum bounds for energy bins
+    mesh :
+        An PyMOAB core instance tagged with all results and
+        relative errors
+    tag_names : iterable
+        Four strs that specify the tag names for the results, relative errors,
+        total results, and relative errors of the total results.
+
+    Notes
+    -----
+    All Mesh/StatMesh attributes are also present via a super() call to
+    StatMesh.__init__().
+
+    """
+
+    def __init__(self, f, tally_number, tag_names=None,
+            mesh_has_mats=False, mc_code='MCNP'):
+        """Create MeshTally object from a filestream open to the second
+        line of a mesh tally header (the neutron/photon line). MeshTally objects
+        should be instantiated through the Meshtal or StatePoint class.
+
+        Parameters
+        ----------
+        f : str or filestream
+            Filestream of the meshtal file or the filename of the state point file.
+        tally_number : int
+            The MCNP fmesh4 tally number (e.g. 4, 14, 24).
+        tag_names : iterable, optional
+            Four strs that specify the tag names for the results, relative
+            errors, total results and relative errors of the total results.
+            This should come from the Meshtal.tags attribute dict.
+        mesh_has_mats : bool
+             If false, Meshtally objects will be created without PyNE material
+             objects.
+        mc_coud : str
+            Monte Carlo code name, could be MCNP or OpenMC.
+        """
+
+        if not HAVE_PYMOAB:
+            raise RuntimeError("PyMOAB is not available, "
+                               "unable to create Meshtally Mesh.")
+
+        self.tally_number = tally_number
+        if mc_code.lower() == 'mcnp':
+            self._read_meshtally_head(f)
+            self._read_column_order(f)
+
+        if tag_names is None:
+            self.tag_names = ("{0}_result".format(self.particle),
+                              "{0}_result_rel_error".format(self.particle),
+                              "{0}_result_total".format(self.particle),
+                              "{0}_result_total_rel_error".format(self.particle))
+        else:
+            self.tag_names = tag_names
+
+        if mc_code.lower() == 'mcnp':
+            self._create_mesh(f, mesh_has_mats)
+
+    def _read_meshtally_head(self, f):
+        """Get the particle type, spacial and energy bounds, and whether or
+        not flux-to-dose conversion factors are being used.
+        """
+        line = f.readline()
+        if ('neutron' in line):
+            self.particle = 'neutron'
+        elif ('photon' in line):
+            self.particle = 'photon'
+
+        # determine if meshtally flux-to-dose conversion factors are being used.
+        line = f.readline()
+        dr_str = 'This mesh tally is modified by a dose response function.'
+        if line.strip() == dr_str:
+            self.dose_response = True
+        else:
+            self.dose_response = False
+
+        # advance the file to the line where x, y, z, bounds start
+        while line.strip() != 'Tally bin boundaries:':
+            line = f.readline()
+
+        self.x_bounds = [float(x) for x in f.readline().split()[2:]]
+        self.y_bounds = [float(x) for x in f.readline().split()[2:]]
+        self.z_bounds = [float(x) for x in f.readline().split()[2:]]
+        # "Energy bin boundaries" contain one more word than "X boundaries"
+        self.e_bounds = [float(x) for x in f.readline().split()[3:]]
+
+        self.dims = [0, 0, 0] + [len(self.x_bounds) - 1,
+                                 len(self.y_bounds) - 1,
+                                 len(self.z_bounds) - 1]
+
+        # skip blank line between enery bin boundaries and table headings
+        f.readline()
+
+    def _read_column_order(self, f):
+        """Create dictionary with table headings as keys and their column
+        location as values. Dictionary is the private attribute _column_idx.
+        """
+        line = f.readline()
+        column_names = line.replace('Rel ', 'Rel_').replace(
+            'Rslt * ', 'Rslt_*_').strip().split()
+        self._column_idx = dict(zip(column_names, range(0, len(column_names))))
+
+    def _create_mesh(self, f, mesh_has_mats):
+        """Instantiate a Mesh object and tag the PyMOAB core instance
+           with results and relative errors.
+        """
+
+        mats = () if mesh_has_mats is True else None
+        super(MeshTally, self).__init__(structured_coords=[self.x_bounds,
+                                                           self.y_bounds, self.z_bounds],
+                                        structured=True, mats=mats)
+
+        num_vol_elements = (len(self.x_bounds)-1) * (len(self.y_bounds)-1)\
+            * (len(self.z_bounds)-1)
+        num_e_groups = len(self.e_bounds)-1
+
+        # get result and relative error data from file
+        result = np.empty(shape=(num_e_groups, num_vol_elements))
+        rel_error = np.empty(shape=(num_e_groups, num_vol_elements))
+        for i in range(0, num_e_groups):
+            result_row = []
+            rel_error_row = []
+            for j in range(0, num_vol_elements):
+                line = f.readline().split()
+                result_row.append(float(line[self._column_idx["Result"]]))
+                rel_error_row.append(
+                    float(line[self._column_idx["Rel_Error"]]))
+
+            result[i] = result_row
+            rel_error[i] = rel_error_row
+
+        # Tag results and error vector to mesh
+        self.tag(self.tag_names[0], tagtype='nat_mesh',
+                 size=num_e_groups, dtype=float)
+        res_tag = self.get_tag(self.tag_names[0])
+        self.tag(self.tag_names[1], tagtype='nat_mesh',
+                 size=num_e_groups, dtype=float)
+        rel_err_tag = self.get_tag(self.tag_names[1])
+
+        if num_e_groups == 1:
+            res_tag[:] = result[0]
+            rel_err_tag[:] = rel_error[0]
+        else:
+            res_tag[:] = result.transpose()
+            rel_err_tag[:] = rel_error.transpose()
+
+        # If "total" data exists (i.e. if there is more than
+        # 1 energy group) get it and tag it onto the mesh.
+        if num_e_groups > 1:
+            result = []
+            rel_error = []
+            for i in range(0, num_vol_elements):
+                line = f.readline().split()
+                result.append(float(line[self._column_idx["Result"]]))
+                rel_error.append(
+                    float(line[self._column_idx["Rel_Error"]]))
+
+            self.tag(self.tag_names[2], size=1,
+                     dtype=float, tagtype='nat_mesh')
+            res_tot_tag = self.get_tag(self.tag_names[2])
+
+            self.tag(self.tag_names[3], size=1,
+                     dtype=float, tagtype='nat_mesh')
+            rel_err_tot_tag = self.get_tag(self.tag_names[3])
+
+            res_tot_tag[:] = result
+            rel_err_tot_tag[:] = rel_error
+
+
 ######################################################
 # private helper functions for structured mesh methods
 ######################################################
